requirements.txt
numpy
pandas
matplotlib
scikit-learn
pytest
jupyter
python-dotenv
ipykernel
